requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2==2.9.3
django-cors-headers==3.10.0
